point프로젝트에서 point조회 하기

+++ src/main/java/spring/work/global/externalApi/workPoint/dto/UserPointInfoApiResponse.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class UserPointInfoApiResponse {
     private String userId;
-    private int point;
+    private Long point;
 }
